// SIMEM-ANG-ADMIN/src/app/store/model/regulatory-classification.model.ts
export interface RegulatoryClassificationModel {
    idConfiguracionClasificacionRegulatoria: string
    codigoDelta: string | null;
    descripcion: string | null;
    fechaCreacion: Date | null;
    deltaFinalAno: number | null;
    deltaFinalDiaMes: number | null;
    deltaFinalDiaSemana: number | null;
    deltaFinalDias: number | null;
    deltaFinalMes: number | null;
    deltaFinalMeses: number | null;
    deltaFinalPeriodo: string | null;
    deltaFinalSemanas: number | null;
    deltaInicialAno: number | null;
    deltaInicialDiaMes: number | null;
    deltaInicialDiaSemana: number | null;
    deltaInicialDias: number | null;
    deltaInicialMes: number | null;
    deltaInicialMeses: number | null;
    deltaInicialPeriodo: string | null;
    deltaInicialSemanas: number | null;    
}